[FBRef] Always group "MP" under "Playing Time"
FBRef is inconsistent in how it displays the "MP" column. For some seasons, it
is displayed as a separate category, while it is grouped under "Playing Time"
for other seasons. This results in a column with NaN values when two seasons are
merged. Therefore, the "MP" column is now always put under "Playing Time".

Closes #79.