[FBref] Drop duplicate season codes
As seasons are converted to a 4 digit season code that ignores
the millennium (e.g., 2012-2013 --> 1213) only post 1923 seasons are supported.
If both a 20xx and 19xx season are available, the 19xx season is now dropped.

In a next major update, we might consider switching to a 6-or 8-digit code, but
for now we use this simple non-breaking fix at the cost of making it impossible
to scrape pre-1923 data.

Fixes #115