[FBref] Informative errors when merge fails (#303)
The FBref scraper attempts to merge data from multiple games, seasons, etc. in a single DataFrame. Sometimes this automatic merge fails, in which case a cryptic error is thrown. Now, a more informative error message is printed listing the additional and/or missing columns and with the suggestion to clear the cache (a common cause is merging new with outdated cached data).

Fixes #161