Handle Selenium 4.13 no longer supporting headless
Selenium does no longer support headless mode for version 4.13 and above. This
was added to the docs and a warning is raised if headless mode is enabled for
a recent version of Selenium.

Also adds a temporary fix for undetected-chromedriver, setting "headless"
explicitly to False, such that soccerdata can still be used with recent
versions of selenium.

Fixes #398