Add option to configure caching max-age via env
This commit adds a "SOCCERDATA_MAXAGE" environment variable. If set, cached data
that is older than the configured max-age will be re-downloaded. This is useful
for testing. We want to check regularly whether the scrapers can still parse the
remote data, but downloading all data in each test run is too slow.