feat(whoscored): option to skip match on error
Adds a "retry_missing" option to the read_events method to control whether to
retry or skip games for which no events are available.

Adds a "skip_on_error" option to the read_events method to control whether to
raise an exception or skip the match when the match page cannot be retrieved.